Remove unnecessary sassc
This reverts 1fd6d41de3c

it was not clear why 1fd6d41de3c was done, but modifying the Gemfile caused this project to become out of sync with the Docker image. Gems no longer installed instantly. This project can be completed in Gitpod without sassc, thus it was removed.